% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{Beautiful graphical representation of multiplication tables}
\usage{
draw(table = 2, modulo = 10, label = FALSE)
}
\arguments{
\item{table}{muliplication table to plot}

\item{modulo}{number of points to use}

\item{label}{show number label}
}
\value{
a ggplot
}
\description{
By placing on a circle 10 points numbered from 1 to 10, and connecting them by a straight line to the point corresponding to its multiplication by 2. (1 must be connected to 1 * 2 = 2, point 2 must be set to 2 * 2 = 4, point 3 to 3 * 2 = 6 and so on). You will obtain an amazing geometric figure that complicates and beautifies itself by varying the number of points and the multiplication table you use.
}
\examples{
draw(table=2,modulo = 10, label=TRUE)
draw(table=2,modulo = 50, label=FALSE)
draw(table=2,modulo = 250)
draw(table=10,modulo = 250)
}
