% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rusk-package.R
\docType{package}
\name{rusk-package}
\alias{rusk-package}
\alias{rusk}
\title{Beautiful graphical representation of multiplication tables}
\description{
Use draw() or draw_app()
}
\author{
vincent <vincent@thinkr.fr>
}
