% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{Beautiful graphical representation of multiplication tables}
\usage{
draw(table = 2, modulo = 10, label = FALSE)
}
\arguments{
\item{table}{muliplication table to plot}

\item{modulo}{number of points to use}

\item{label}{show number label}
}
\value{
a ggplot
}
\description{
Beautiful graphical representation of multiplication tables
}
\examples{
draw(table=2,modulo = 10, label=TRUE)
draw(table=2,modulo = 50, label=FALSE)
draw(table=2,modulo = 250)
draw(table=10,modulo = 250)
}
