% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{Beautiful graphical representation of multiplication tables}
\usage{
draw(table = 2, modulo = 10)
}
\arguments{
\item{table}{which muliplication table to plot}

\item{modulo}{number of points to use}
}
\value{
a ggplot
}
\description{
Beautiful graphical representation of multiplication tables
}
\examples{
draw(table=2,modulo = 50)
draw(table=2,modulo = 250)
draw(table=10,modulo = 250)
}
