% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rusk-package.R
\docType{package}
\name{rusk-package}
\alias{rusk-package}
\alias{rusk}
\title{Beautiful graphical representation of multiplication tables}
\description{
By placing on a circle 10 points numbered from 1 to 10, and connecting them by a straight line to the point corresponding to its multiplication by 2. (1 must be connected to 1 * 2 = 2, point 2 must be set to 2 * 2 = 4, point 3 to 3 * 2 = 6 and so on). You will obtain an amazing geometric figure that complicates and beautifies itself by varying the number of points and the multiplication table you use.
}
\details{
Use draw() or draw_app()
}
\references{
\url{https://www.youtube.com/embed/qhbuKbxJsk8?rel=0}

\url{https://www.youtube.com/embed/-X49VQgi86E?rel=0}
}
\author{
vincent <vincent@thinkr.fr>
}
